rename files, move some methods
