+++ htsget_server/server.py
@@ -8,7 +8,7 @@ from config import PORT, DEBUG_MODE
 app = connexion.App(__name__, specification_dir='./')
 CORS(app.app)
 
-app.add_api('openapi.yml', pythonic_params=True, strict_validation=True)
+app.add_api('htsget_openapi.yaml', pythonic_params=True, strict_validation=True)
 app.add_api('drs_openapi.yaml', pythonic_params=True, strict_validation=True)
 
 # Just leaving this here as a note: these are all of the pythonic params that
